Build test cases for folio management
While building those test cases for folios, I also fix some backend bugs. Those bugs are commented out in the original file and could be reproduced by remove those comment symbols. The reason for this is we may want to revert back in the future and this makes reverting back easier.